move run.py to be the _main__
